Support an EEPROM-based keymap overlay & Focus
This adds the necessary code to support having a keymap overlay in EEPROM (with
fallback to PROGMEM). It also adds support for Focus, so we can change the
contents of EEPROM more easily.

The Focus parts enable a few hooks that are not (currently) required or used by
Chrysalis, but which are useful nevertheless, and can be built upon.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>